% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_search.R
\name{apply_lambda_to_ridge}
\alias{apply_lambda_to_ridge}
\title{Functions to help searching lambda in ridge VPA
Pass multiple lambda to \code{run_ridge_vpa}}
\usage{
apply_lambda_to_ridge(lambda, sigma_constraint, fcurrent_year, ...)
}
\arguments{
\item{lambda}{Vector of multiple lambdas}

\item{sigma_constraint}{\code{sigma.constraint} in \code{\link[frasyr]{vpa}}}

\item{fcurrent_year}{\code{fc.year} in \code{\link[frasyr]{vpa}}}

\item{...}{Other args to pass \code{\link[frasyr]{vpa}}}
}
\description{
This function is a component of \code{\link{make_rhotable}}
}
