% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logabs}
\alias{logabs}
\title{Convert value to logarithm of absolute}
\usage{
logabs(x, convert)
}
\arguments{
\item{x}{The value to be converted}

\item{convert}{If FALSE, x will be returned without conversion}
}
\description{
Convert value to logarithm of absolute
}
