% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_search.R
\name{make_rhotable}
\alias{make_rhotable}
\title{Make Mohn's rho table}
\usage{
make_rhotable(lambda, sigma_constraint, fcurrent_year, ...)
}
\arguments{
\item{lambda}{Vector of multiple lambdas}

\item{sigma_constraint}{\code{sigma.constraint} in \code{\link[frasyr]{vpa}}}

\item{fcurrent_year}{\code{fc.year} in \code{\link[frasyr]{vpa}}}

\item{...}{Other args to pass \code{\link[frasyr]{vpa}}}
}
\description{
Make Mohn's rho table
}
