% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_search.R
\name{plot_rhos}
\alias{plot_rhos}
\title{Plot Mohn's rho table for visual check}
\usage{
plot_rhos(rhotable_df, log_abs = TRUE)
}
\arguments{
\item{rhotable_df}{Table of Mohn's rho in data frame format}

\item{log_abs}{If TRUE, plot logarithm of absolute value}
}
\description{
Actual decision process should be automated!
}
