% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_vpa.R
\name{pull_mohns_rho}
\alias{pull_mohns_rho}
\title{Pull Mohn's rho from output of ridge VPA}
\usage{
pull_mohns_rho(result_ridge_vpa)
}
\arguments{
\item{result_ridge_vpa}{Output of \code{link[frasyr]{vpa}}}
}
\description{
Pull Mohn's rho from output of ridge VPA
}
