% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_vpa.R
\name{run_ridge_vpa}
\alias{run_ridge_vpa}
\title{Run ridge VPA}
\usage{
run_ridge_vpa(sigma_constraint, lambda, dat, fcurrent_year, ...)
}
\arguments{
\item{sigma_constraint}{Equals to \code{sigma.const} in \code{\link[frasyr]{vpa}}}

\item{lambda}{ridge回帰係数}

\item{dat}{data for vpa}

\item{fcurrent_year}{Equals to \code{fc.year} in \code{\link[frasyr]{vpa}}}

\item{...}{Other arguments to pass to \code{\link[frasyr]{vpa}}}
}
\description{
Run ridge VPA
}
