% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{select_good_lambda}
\alias{select_good_lambda}
\title{Select good lambda with minimum absolute of Mohn's rho of given variable}
\usage{
select_good_lambda(mohns_rho_tbl, select_by_retro_bias_of, n_returned)
}
\arguments{
\item{mohns_rho_tbl}{Output of \code{\link{make_rhotable}}}

\item{select_by_retro_bias_of}{Character string of variable one of
\describe{
  \item{"N"}{number of fishes}
  \item{"B"}{biomass}
  \item{"SSB"}{spawning stock biomass}
  \item{"R"}{recruitment}
  \item{"F"}{fishing mortality coefficient}
}}

\item{n_returned}{Number of lambdas to be returned. \code{n_returned = 2} is
appropriate for finding candidates in the first run with large step,
and \code{n_returned} is for deciding the best value in the final run with
small step.}
}
\description{
Select good lambda with minimum absolute of Mohn's rho of given variable
}
