% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_search.R
\name{tidy_rhotable}
\alias{tidy_rhotable}
\title{Convert Mohn's rho table to data frame}
\usage{
tidy_rhotable(rhotable, lambda)
}
\arguments{
\item{rhotable}{Output of \code{\link{apply_lambda_to_ridge}}}

\item{lambda}{Vector of multiple lambdas}
}
\description{
This function is a component of \code{\link{make_rhotable}}
}
